feat: switches from username model to ids

+++ src/main/kotlin/com/rchat/server/controllers/ChatController.kt
@@ -13,6 +13,8 @@ import org.springframework.stereotype.Controller
 import com.fasterxml.jackson.annotation.JsonView
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.messaging.Message
+import org.springframework.messaging.handler.annotation.support.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PathVariable
 import java.time.LocalDate
@@ -27,18 +29,18 @@ class ChatController(
 ) {
     @Transactional
     @JsonView(View.AllWithId::class)
-    @MessageMapping("/user/{recipient}/{sender}/")
-    @SendTo("/chatTopic/{recipient}/", "/chatTopic/{sender}/")
+    @MessageMapping("/user/{recipientID}/{senderID}/")
+    @SendTo("/chatTopic/{recipientID}/", "/chatTopic/{senderID}/")
     fun processPersonal(
-        @DestinationVariable recipient: String,
-        @DestinationVariable sender: String,
+        @DestinationVariable recipientID: String,
+        @DestinationVariable senderID: String,
         msg: String
     ): PersonalMessage {
         val date = LocalDate.now()
         val time = LocalTime.now()
         val message = PersonalMessage(
-            userService.getByName(sender),
-            userService.getByName(recipient),
+            userService.getById(senderID.toInt()),
+            userService.getById(recipientID.toInt()),
             time,
             date,
             msg
@@ -89,7 +91,9 @@ class ChatController(
         @DestinationVariable sender: String,
         msgId: String
     ): Map<String, String> {
-//        val message = personalMessageRepo.getById(msgId.toInt())
+        val message = personalMessageRepo.getById(msgId.toInt())
+        if (message.recipient == userService.getByName(recipient) && message.sender == userService.getByName(sender))
+            return mapOf("error" to "access denied")
         personalMessageRepo.deletePersonalMessageById(msgId.toInt())
         return mapOf("deleted" to msgId)
     }
